Changed Loading Listings Behavior
Essentially, I made it so when changing a filter behavior or edit the
Search it takes the preloaded listings and filters those instead of
querying the database constantly.
Also I randomly included a none before for like wish lists, so I made
the condition filter bypass the nones so they are all still shown!